Added Registration tests in original test file
Shifted Registration Tests from another file into this.
